requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
easyocr>=1.7.0
layoutparser>=0.3.4

# Speech Processing
openai-whisper>=20230314
librosa>=0.10.0
soundfile>=0.12.0

# NLP
sentencepiece>=0.1.99
sacremoses>=0.0.53
nltk>=3.8

# API & Serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# MLOps
mlflow>=2.5.0
dvc>=3.0.0
wandb>=0.15.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Monitoring
prometheus-client>=0.17.0
evidently>=0.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Development
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
black>=23.7.0
flake8>=6.0.0
